cc-3683: rendering scripts from editor/display templates
The proper solution to a problem like this is to use RenderSection("sectionName", false) in conjunction with @section sectionName. However, that works for partial views only (AFAIK). The FormUpload.cshtml file is an Editor Template, hence any @sections defined in the Editor Templates are ignored.

Solution as described here: http://stackoverflow.com/a/5433722/1042643 (copy/pasted code)