cc-4150: Changed session renewal prompt time
Showing prompt 3 minutes before expiration, but giving user 2 minutes to respond (to account for server round trip time). For example, if the session expire at 2:03 pm, the user will be prompted to renew their session at 2:00 pm. if the user clicks OK to renew the session at 2:01:59 pm (i.e., with 1 second remaining) the additional minute allows for a round trip to the server to renew the user session.